flight booking simulation using dynamic pricing/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.18
psycopg2-binary==2.9.7
pydantic==1.10.12
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.24.1
apscheduler==3.10.1
reportlab==4.0.0
